+++ flexmeasures/cli/data_add.py
@@ -37,6 +37,7 @@ from flexmeasures.data.models.data_sources import (
 )
 from flexmeasures.utils import flexmeasures_inflection
 from flexmeasures.utils.time_utils import server_now
+from flexmeasures.utils.unit_utils import convert_units
 
 
 @click.group("add")
@@ -320,6 +321,14 @@ def add_initial_structure():
     help="Source of the beliefs (an existing source id or name, or a new name).",
 )
 @click.option(
+    "--unit",
+    required=False,
+    type=str,
+    help="Unit of the data, for conversion to the sensor unit, if possible (a string unit such as 'kW' or 'm³/h').\n"
+    "Hint: to switch the sign of the data, prepend a minus sign.\n"
+    "For example, when assigning kW consumption data to a kW production sensor, use '-kW'.",
+)
+@click.option(
     "--horizon",
     required=False,
     type=int,
@@ -400,6 +409,7 @@ def add_beliefs(
     file: str,
     sensor_id: int,
     source: str,
+    unit: Optional[str] = None,
     horizon: Optional[int] = None,
     cp: Optional[float] = None,
     resample: bool = True,
@@ -470,6 +480,13 @@ def add_beliefs(
         parse_dates=True,
         **kwargs,
     )
+    if unit is not None:
+        bdf["event_value"] = convert_units(
+            bdf["event_value"],
+            from_unit=unit,
+            to_unit=sensor.unit,
+            event_resolution=sensor.event_resolution,
+        )
     try:
         TimedBelief.add(
             bdf,
@@ -560,15 +577,11 @@ def add_holidays(
 @fm_add_data.command("forecasts")
 @with_appcontext
 @click.option(
-    "--asset-type",
-    type=click.Choice(["Asset", "Market", "WeatherSensor"]),
-    help="The generic asset type for which to generate forecasts."
-    " Follow up with Asset, Market or WeatherSensor.",
-)
-@click.option(
-    "--asset-id",
-    help="Populate (time series) data for a single asset only. Follow up with the asset's ID. "
-    "We still need --asset-type, as well, so we know where to look this ID up.",
+    "--sensor-id",
+    "sensor_ids",
+    multiple=True,
+    required=True,
+    help="Create forecasts for this sensor. Follow up with the sensor's ID. This argument can be given multiple times.",
 )
 @click.option(
     "--from-date",
@@ -585,7 +598,7 @@ def add_holidays(
 @click.option(
     "--resolution",
     type=int,
-    help="Resolution of forecast in minutes. If not set, resolution is determined from the asset to be forecasted",
+    help="Resolution of forecast in minutes. If not set, resolution is determined from the sensor to be forecasted",
 )
 @click.option(
     "--horizon",
@@ -603,8 +616,7 @@ def add_holidays(
     " config settings to that of the remote server. To process the job, run a worker to process the forecasting queue. Defaults to False.",
 )
 def create_forecasts(
-    asset_type: str = None,
-    asset_id: int = None,
+    sensor_ids: List[int],
     from_date_str: str = "2015-02-08",
     to_date_str: str = "2015-12-31",
     horizons_as_hours: List[str] = ["1"],
@@ -616,10 +628,10 @@ def create_forecasts(
 
     For example:
 
-        --from_date 2015-02-02 --to_date 2015-02-04 --horizon_hours 6
+        --from_date 2015-02-02 --to_date 2015-02-04 --horizon_hours 6 --sensor-id 12 --sensor-id 14
 
         This will create forecast values from 0am on May 2nd to 0am on May 5th,
-        based on a 6 hour horizon.
+        based on a 6-hour horizon, for sensors 12 and 14.
 
     """
     # make horizons
@@ -639,26 +651,26 @@ def create_forecasts(
         event_resolution = None
 
     if as_job:
-        for horizon in horizons:
-            # Note that this time period refers to the period of events we are forecasting, while in create_forecasting_jobs
-            # the time period refers to the period of belief_times, therefore we are subtracting the horizon.
-            create_forecasting_jobs(
-                old_sensor_id=asset_id,
-                horizons=[horizon],
-                start_of_roll=forecast_start - horizon,
-                end_of_roll=forecast_end - horizon,
-            )
+        for sensor_id in sensor_ids:
+            for horizon in horizons:
+                # Note that this time period refers to the period of events we are forecasting, while in create_forecasting_jobs
+                # the time period refers to the period of belief_times, therefore we are subtracting the horizon.
+                create_forecasting_jobs(
+                    sensor_id=sensor_id,
+                    horizons=[horizon],
+                    start_of_roll=forecast_start - horizon,
+                    end_of_roll=forecast_end - horizon,
+                )
     else:
         from flexmeasures.data.scripts.data_gen import populate_time_series_forecasts
 
         populate_time_series_forecasts(
             db=app.db,
+            sensor_ids=sensor_ids,
             horizons=horizons,
             forecast_start=forecast_start,
             forecast_end=forecast_end,
             event_resolution=event_resolution,
-            old_sensor_class_name=asset_type,
-            old_sensor_id=asset_id,
         )